requirements.txt
clickhouse-driver
clickhouse_connect
langchain
langchain-core
langgraph
pydantic
python-dotenv
pandas